+++ platform/windows/cs/Adguard.Dns/SolutionInfo.cs
@@ -12,5 +12,5 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCopyright("Copyright © Adguard Software Ltd 2009-2024")]
 [assembly: AssemblyTrademark("Copyright © Adguard Software Ltd 2024")]
 [assembly: AssemblyCulture("")]
-[assembly: AssemblyVersion("2.8.31")]
-[assembly: AssemblyFileVersion("2.8.31")]
+[assembly: AssemblyVersion("2.8.32")]
+[assembly: AssemblyFileVersion("2.8.32")]